LibBareMetal: Accept a u16 type pointer in the appropriate IO functions

+++ Libraries/LibBareMetal/IO.h
@@ -64,7 +64,7 @@ inline u32 in32(u16 port)
     return value;
 }
 
-inline void repeated_in16(u16 port, u8* buffer, int buffer_size)
+inline void repeated_in16(u16 port, u16* buffer, int buffer_size)
 {
     asm volatile("rep insw"
                  : "+D"(buffer), "+c"(buffer_size)
@@ -87,7 +87,7 @@ inline void out32(u16 port, u32 value)
     asm volatile("outl %0, %1" ::"a"(value), "Nd"(port));
 }
 
-inline void repeated_out16(u16 port, const u8* data, int data_size)
+inline void repeated_out16(u16 port, const u16* data, int data_size)
 {
     asm volatile("rep outsw"
                  : "+S"(data), "+c"(data_size)
